Keep server-only close-rule code out of client bundles
validateTicketClosure.ts imports hasPermission from @alga-psa/auth, whose
barrel transitively pulls nextAuthOptions -> node:crypto/async_hooks/buffer.
Because the @alga-psa/tickets/lib barrel re-exported validateTicketClosure,
the client component TicketList.tsx (which imports that barrel for unrelated
helpers) dragged the server-only chain into the browser bundle, breaking the
EE server build (ee-workflows-build-guard) and fresh-install-e2e.

Split the client-safe pieces (CloseRule* types, CLOSE_RULE_REQUIRED_FIELDS,
CLOSE_RULE_REQUIRED_FIELD_LABELS, TicketCloseValidationError) into a new
closeRuleConstants.ts that has only type imports. The barrel now re-exports
those from closeRuleConstants and no longer references validateTicketClosure;
the server-only enforce/evaluate functions stay in validateTicketClosure.ts
(re-exported there for back-compat) and are reached via the deep path. Pointed
the one barrel consumer of enforceTicketCloseRules (client-portal
client-tickets.ts, a server action) at @alga-psa/tickets/lib/validateTicketClosure.

tickets + client-portal typecheck clean.
